+++ packages/client/hmi-client/src/page/data-explorer/components/preview-panel.vue
@@ -20,7 +20,7 @@
 					:highlight="searchTerm"
 					:is-editable="false"
 				/>
-				<dataset
+				<teraDataset
 					v-else-if="previewItemResourceType === ResourceType.DATASET"
 					:asset-id="previewItemId"
 					:project="resources.activeProject"
@@ -60,7 +60,7 @@ import useResourcesStore from '@/stores/resources';
 import { ResultType, ResourceType } from '@/types/common';
 import { isDocument } from '@/utils/data-util';
 import Document from '@/components/documents/Document.vue';
-import Dataset from '@/components/dataset/Dataset.vue';
+import teraDataset from '@/components/dataset/tera-dataset.vue';
 import Model from '@/components/models/Model.vue';
 import TeraSlider from '@/components/widgets/tera-slider.vue';
